requirements.txt
fastapi
uvicorn
tensorflow
pillow
redis